// src/interfaces/Product.interface.ts
export interface IProduct {
  id?: string;
  title?: string;
  sub_title?: string;
  description?: string;
  cost_price?: string;
  product_status?: string;
  color?: string;
  size?: string;
  product_availability?: string;
  category_id?: string;
  selling_price?: string;
  discount?: string;
  discount_price?: string;
  store_Name?: string;
  review?: string;
  rating?: string;
  sold?: string;
}